add lib to approximate Gamma distribution mixtures
Purpose:

gamma_approx offers routines to fit a single
Gamma distribution to approximate some given
mixture of Gamma distributions. The parameters
of the approximating Gamma distribution are
chosen to minimise the KL-divergence from the
approximation to the original Gamma mixture.

For each fit, the algorithm is:

1.	Compute the expected rate and expected
	log rate of the input Gamma mixture.

2.	Obtain a two-dimensional system of
	nonlinear constraints equating the
	expected rate and expected log rate
	of the approximation with the
	expected rate and expected log rate
	of the input mixture, respectively.

3.	Compute the optimal value of the
	approximating shape parameter by
	solving a one-dimensional problem
	to invert the function

	y = digamma(x) - log(x).

	This is perfomed numerically using
	Halley's method and series
	approximations of digamma and its
	first two derivatives.

4.	Solve for the optimal value of
	the rate parameter in terms of the
	optimal shape parameter and the
	input data. In constrast to step
	3, this is immediate.

Note that performance appears to be greatly
accelerated when solving large batches of
these approximation problems. There must be
considerable overhead in the python/cython
wrapper bindings. This could use some more
investigation.

The batched implementation can fit 320k mixtures
of size 1 -- 20 component Gammas in less than
0.175 seconds on my machine, using one core.
That's a running time of 540 nanoseconds per fit.