define fixed gamma hmm to benchmark against
Fixed Gamma HMM aka fghmm takes a very simple
approach to modelling uncertainty about the
noise rate. It roughly works like so:

1.	Pick an integer r >= 1
2.	Fix r different predefined noise rates,
	spread out across the plausible range
	defined by the prior distribution on the
	noise rate
3.	Make r copies of the base HMM
4.	Attach the i-th predefined noise rate to
	the i-th copy of the base HMM. Fix the
	noise rate so it cannot be modified by
	new information.
5.	Combine the resulting r HMMs together with
	a disjoint union to form a single HMM of
	r HMMs with different noise rates that
	evolve independently.

The last step isn't strictly necessary -- the r
different models with different noise rates could
be updated independently, and then care would need
to be taken with normalisation to ensure their
log-probs are comparable. It is about the same
amount of computational work to combine them into
a single HMM.

Note that we've implemented a variation of step 2,
instead of fixing a single noise rate we fix a
reasonably concentrated Gamma distribution for the
noise rate concentrated on some assumed noise level.
This is an implementation detail that lets us reuse
our existing code, but it also expresses that our
knowledge of the noise level is not certain within
each of these r alternaitve models.